requirements.txt
requests>=2.31.0
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
geopandas>=0.14.0
shapely>=2.0.0

